services/api/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
mlflow==2.9.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
lightfm==1.17
prophet==1.1.5
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
python-multipart==0.0.6
httpx==0.25.2
aioredis==2.0.1
